Update ExrToCameraRaw.py
This fix the FOV computation from the projection matrix.

There is a attempt to get the clip planes too, but the line equations are not right somehow. The code is still there, if someone want to fix it...

It won't work with anamorphic images. The pixel ratio is embedded in the metadata so it's possible to handle that case, if someone want to add support for that...